+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.11.06 (pending)
 * add a convenience method for setting the content type
+* add a common reference to popular mime types
 
 ## 3.11.05
 *  issue #383 some problems with relocated packages.