+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 3.11.06 (pending)
+## 3.11.07 (pending)
+* add a new object contructor on JSONOBject for basic object serialization
+
+## 3.11.06
 * add a convenience method for setting the content type
 * add a common reference to popular mime types
 * cache methods on Config were not returning the config for the builder pattern.